requirements.txt
streamlit==1.24.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
python-pptx==0.6.23
sentence-transformers==2.2.2
faiss-cpu==1.7.2
transformers==4.37.0
torch==2.1.0